requirements.txt
discord.py>=2.3.2
fastapi>=0.110.0
uvicorn[standard]>=0.23.2
python-dotenv>=1.0.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.0.0
